Extended max length to be a long allowing very big files

+++ src/main/java/de/s42/dl/services/DLParameter.java
@@ -46,7 +46,7 @@ public @interface DLParameter
 
 	public String value();
 
-	public int maxLength() default 1024;
+	public long maxLength() default 1024L;
 
 	public String pattern() default "";
